requirements.txt
# absl-py==0.10.0
# asgiref==3.2.10
# astroid==2.4.2
# astunparse==1.6.3
# azure-cognitiveservices-speech==1.14.0
# backcall==0.2.0
# beautifulsoup4==4.6.0
# bs4==0.0.1
# cached-property==1.5.2
# cachetools==4.1.1
# certifi==2020.6.20
# chardet==3.0.4
# colorama==0.4.4
# commonmark==0.9.1
# decorator==4.4.2
# dictdiffer==0.8.1
# Django
# django-extensions==3.0.9
# djangorestframework==3.12.1
# gast==0.3.3
# gensim==3.8.3
# google-api-core==1.23.0
# google-api-python-client==1.12.5
# google-auth==1.22.1
# google-auth-httplib2==0.0.4
# google-auth-oauthlib==0.4.1
# google-pasta==0.2.0
# googleapis-common-protos==1.52.0
# grpcio==1.32.0
# h5py==2.10.0
# httplib2==0.18.1
# idna==2.10
# importlib-metadata==2.0.0
# ipykernel==5.3.4
# ipython
# ipython-genutils==0.2.0
# isort
# jedi==0.17.2
# joblib
# JPype1==1.1.2
# JPype1-py3==0.5.5.4
# jupyter-client==6.1.7
# jupyter-core==4.6.3
# Keras==2.4.3
# keras-bert==0.86.0
# keras-embed-sim==0.8.0
# keras-layer-normalization==0.14.0
# keras-multi-head==0.27.0
# keras-pos-embd==0.11.0
# keras-position-wise-feed-forward==0.6.0
# Keras-Preprocessing==1.1.2
# keras-self-attention==0.46.0
# keras-transformer==0.38.0
# konlpy==0.5.2
# lazy-object-proxy==1.4.3
# lexrankr==0.1.6
# lxml==4.6.1
# Markdown
# mccabe==0.6.1
# networkx==2.5
# notion==0.0.25
# numpy==1.18.5
# oauthlib==3.1.0
# opt-einsum==3.3.0
# pandas==1.1.3
# parso==0.7.1
# pexpect==4.8.0
# pickleshare==0.7.5
# prompt-toolkit==3.0.8
# protobuf==3.13.0
# ptyprocess==0.6.0
# pyasn1==0.4.8
# pyasn1-modules==0.2.8
# Pygments==2.7.1
# pylint==2.6.0
# PySocks==1.7.1
# python-dateutil==2.8.1
# python-slugify==4.0.1
# pytz==2020.1
# PyYAML==5.3.1
# pyzmq==19.0.2
# requests==2.24.0
# requests-oauthlib==1.3.0
# rsa==4.6
# scikit-learn==0.23.2
# scipy==1.5.2
# six==1.15.0
# sklearn==0.0
# smart-open==3.0.0
# soupsieve==2.0.1
# sqlparse==0.4.1
# tensorboard==2.3.0
# tensorboard-plugin-wit==1.7.0
# tensorflow==2.3.1
# tensorflow-estimator==2.3.0
# termcolor==1.1.0
# text-unidecode==1.3
# threadpoolctl==2.1.0
# toml==0.10.1
# tornado==6.0.4
# traitlets==4.3.3
# tweepy==3.9.0
# typed-ast==1.4.1
# typing-extensions==3.7.4.3
# tzlocal==2.1
# uritemplate==3.0.1
# uritemplate.py==3.0.2
# urllib3==1.25.10
# wcwidth==0.2.5
# Werkzeug==1.0.1
# wrapt==1.12.1
# zipp==3.3.0
lexrankr
django
azure-cognitiveservices-speech
google-api-core
google-api-python-client
google-auth
google-auth-httplib2
google-auth-oauthlib
google-pasta
googleapis-common-protos
notion
bert-extractive-summarizer
transformers
torch